Rename `Common` to `OpenSaveGame`
This relates to the only functionality that is left in `Common.h`.
