+++ appOPHD/States/MainMenuState.cpp
@@ -4,7 +4,7 @@
 #include "PlanetSelectState.h"
 
 #include "../Cache.h"
-#include "../Common.h"
+#include "../OpenSaveGame.h"
 #include "../Constants/Strings.h"
 #include "../Constants/UiConstants.h"
 #include "../ShellOpenPath.h"
