+++ appOPHD/States/MapViewStateIO.cpp
@@ -9,7 +9,7 @@
 #include "Route.h"
 
 #include "../Cache.h"
-#include "../Common.h"
+#include "../OpenSaveGame.h"
 #include "../MeanSolarDistance.h"
 #include "../Constants/Strings.h"
 #include "../IOHelper.h"
